Don't crash here if pid is None.
git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@2844 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: f8955336328e2b0fbeab7c7969d0d0dd740a04aa

+++ tv/platform/gtk-x11/frontend_implementation/UIBackendDelegate.py
@@ -209,6 +209,8 @@ def ShowTextEntryDialogAsync(title, description, buttons, default, prefillCallba
     gtkDialog.show()
 
 def pidIsRunning(pid):
+    if pid is None:
+        return False
     try:
         os.kill(pid, 0)
         return True
@@ -324,6 +326,8 @@ class UIBackendDelegate:
         primary.set_text(text)
 
     def killDownloadDaemon(self, oldpid):
+        if oldpid is None:
+            return
         if pidIsRunning(oldpid):
             try:
                 os.kill(oldpid, signal.SIGTERM)
